Wrapper around `SEXP` so that we can make it Send / Sync
