+++ build.rs
@@ -452,6 +452,7 @@ fn generate_bindings(r_paths: &InstallationPaths, version_info: &RVersionInfo) {
     // to bindgen, and lets you build up options for
     // the resulting bindings.
     let mut bindgen_builder = bindgen::Builder::default()
+        .new_type_alias_deref("SEXP")
         .emit_diagnostics()
         .allowlist_function(&allowlist_pattern)
         .allowlist_var(&allowlist_pattern)
